Parse SPICE Extraction feedback and convert into the KLayout database format
## Steps

* `Magic.SpiceExtraction`
  * A `feedback.xml` is now created, with the contents being the SPICE Extraction feedback in the KLayout database format

## Misc Enhancements/Bugfixes
* `%OL_METRICS_F` now uses Decimals instead of Floats
* Fixed crash when attempting to dump a Decimal of infinite value

+++ openlane/steps/step.py
@@ -171,11 +171,11 @@ class DefaultOutputProcessor(OutputProcessor[Dict[str, Any]]):
             self.current_rpt = None
         elif line.startswith(METRIC_LOCUS):
             command, name, value = line.split(" ", maxsplit=3)
-            metric_type: Union[Type[str], Type[int], Type[float]] = str
+            metric_type: Union[Type[str], Type[int], Type[Decimal]] = str
             if command.endswith("_I"):
                 metric_type = int
             elif command.endswith("_F"):
-                metric_type = float
+                metric_type = Decimal
             self.generated_metrics[name] = metric_type(value)
         elif self.current_rpt is not None:
             # No echo- the timing reports especially can be very large
